VaultConsul-Labs/appV1/requirements.txt
flask
hvac
pika
pymongo
gunicorn
flask-cors
python-consul
flask-restplus